fp: full website update
I decided to put the lecture code in a new folder so i added a smart constructor for that. We also have a couple of lectures in a different place so I added a case for that. I also changed some colours. Finally, I added a smart constructor for min sheet links